refactor: address code-style audit findings
Audit driven by CODE_STYTLE_RUST.md.

Bad practice fixes:
- Replace 'e.to_string() == "BucketNotEmpty"' sentinel-string pattern in
  s3impl/grpc with io::ErrorKind::DirectoryNotEmpty round-trip.
- Same treatment for the MalformedXML carrier in put_bucket_versioning
  using io::ErrorKind::InvalidData.
- Drop unused indicatif helpers (transfer_bar, human_bytes) from
  progress.rs and unwrap()s on ProgressStyle templates, replacing the
  remaining one with .expect("static template literal must parse").

#[allow] hygiene (style guide: every suppression needs a real reason):
- Remove redundant '#[allow(clippy::missing_errors_doc)]' impl
  annotations — the lint is globally disabled in .cargo/config.toml.
- Add justification comments to every remaining cast/wildcard/dead-code
  allow across cli, grpc, s3impl, storage.

File budget:
- src/cli/serve.rs 896 → 707 lines by extracting admin endpoint handlers
  to src/cli/serve_admin.rs (198 lines). The routing in AdminService
  stays put; only the per-endpoint handlers and their DTOs moved.
